third_party/WebKit/LayoutTests/media/track/track-node-add-remove.html

<!DOCTYPE html>
<title>Tests adding and removing track node.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function() {
  function assert_track_list(trackList) {
    assert_equals(video.textTracks.length, trackList.length);
    for (var i = 0; i < trackList.length; i++)
      assert_equals(video.textTracks[i], trackList[i].track);
  }

  var video = document.createElement('video');
  var tracka = document.createElement('track');
  video.appendChild(tracka);
  var trackb = document.createElement('track');
  video.appendChild(trackb);

  // Adding tracks outside the DOM tree.
  assert_track_list([tracka, trackb]);

  // Inserting the parent video element into the document.
  document.body.appendChild(video);
  assert_track_list([tracka, trackb]);

  // Inserting and removing another track in the document.
  var trackc = document.createElement('track');
  video.appendChild(trackc);
  assert_track_list([tracka, trackb, trackc]);

  trackb.parentNode.removeChild(trackb);
  assert_track_list([tracka, trackc]);

  // Removing the video from the document.
  document.body.removeChild(video);
  assert_track_list([tracka, trackc]);

  tracka.parentNode.removeChild(tracka);
  assert_track_list([trackc]);
});
</script>
